III-3949 - Move Calendar & TimeSpan to separate file
